+++ README.md
@@ -1,2 +1,41 @@
-## 高版本的Python会有ssl的问题
-pip3 install --trusted-host pypi.org --upgrade baostock
+# 项目 TODO list
+- 1.各组员自行学习量化知识点，拉取当前代码库。
+- 2.[量化交易笔记.md](%E9%87%8F%E5%8C%96%E4%BA%A4%E6%98%93%E7%AC%94%E8%AE%B0.md) 是我自己整理的一些相关笔记，供大家参考。同时后面的计划和思考，也是基于这个笔记进行的。
+- 3.选择一个包含机器学习的策略方向，获取数据集，建立初版模型。
+  - 思考：这里的实现，需要由1~2个人来做，因为工作量比较大，由1个人负责整体思路构想，另1个人帮忙找数据来源，以及处理数据，比如清洗等。
+  - ***论文点***：为什么要选择这个策略？优劣势分析
+  - ***论文点***：建模，数据获取的方式、预处理、数据清洗，建立特种工程的过程和方式，特征工程中，选择特征因子的原因。
+  - ***论文点***：编程实现的方式，实现细节等。
+  - [ ] 负责人1：
+    - [ ] 代码已完成：
+    - [ ] PPT已完成：
+    - [ ] 论文已完成：
+  - [ ] 负责人2：
+    - [ ] 代码已完成：
+    - [ ] PPT已完成：
+    - [ ] 论文已完成：
+- 4.模型调优，可包括：超参数调优，过拟合处理，可视化分析等，如果能有能力进行集成学习，那更好了。
+  - 思考：这里的调优，可以由1个，或多个人来做，选择不同的调优方向，选择任务时不用着急确定具体优化点，可以先确定一个大方向，比如超参数调优，过拟合处理等。具体的优化点可以看模型运行的结果，和后续的讨论中确定。
+  - ***论文点***：如何优化参数？通过xx方式，优化了xx，达到了xx样的效果。
+  - [ ] 负责人3：
+    - [ ] 代码已完成：
+    - [ ] PPT已完成：
+    - [ ] 论文已完成：
+  - [ ] 负责人4：
+    - [ ] 代码已完成：
+    - [ ] PPT已完成：
+    - [ ] 论文已完成：
+- 5.风险评估
+  - ***论文点***：如何计算风险指标？如何证明风险可控？是高风险还是低风险。
+  - 负责人5：
+  - [ ] 代码已完成：
+  - [ ] PPT已完成：
+  - [ ] 论文已完成：
+
+## 另外：
+  - 各组员根据自己感兴趣的研究方向，选择某个模块，填写负责人，在相应的checkbox中打钩，后提交PR合并。
+    - 例如项目的初模型的开发，由我来做，那我就在负责人这里填上的我的名字，然后提交PR
+  - 每人根据自己的研究方向，需要同时生成两页PPT，完成并上传代码库后，在相应的checkbox中打钩。
+  - baostock是老师提供的免费因子库，[baostock_demo.py](demo/baostock_demo.py) 是一个简单的demo，演示了如何使用获取数据。
+
+# 注意！！！在项目正式开始开发之后，所有人每次要修改、提交PR之前，一定要先pull一下最新的代码，然后把最近的改动（如果有）合并到自己的分支，再修改/提交，避免冲突。
