+++ 大作业作业要求.md
@@ -1,0 +1,3 @@
+- 大作业作业要求
+  - 1.小组作业：小组自选主题，用深度学习的方法解决（自行找数据集，设计训练模型），五一假期结束后（5.10）的课上进行小组ppt汇报；
+  - 2.个人作业：围绕小组的项目背景，结合你做的细分方向，写一篇2000-3000字左右的论文。
